Add pseudo folder rendering
